Perfecting the Horizontal Landing
Big changes:
HORIZONTAL_LANDING
- Ran a simulation with different constant TWR ships at different
starting orbits to determine what to set the periapsis of the landing
orbit. Then I added a "Fuel Factor" because the ships lose mass so it
will lower the periapsis a little bit to take that into account.
- I added a Gravity Turn Factor because the horizontal acceleration
drops off due to the fact that the ship turns downward because of the
funciton of a gravity turn.
- Changed the cutoff function to a calculated value based on the time it
takes to free fall to the ground and its current horizontal speed. That
way if it is okay to just drop down I don't have to be at 0 speed
horizontally and 0 distance to the spot. Works pretty well
- Changed it so that when the ship reaches the landing perigee it will
be over the landing spot
- Left the Buffer Distance in there in case it's needed but currently it
is set to 0, works well as is
VERTICAL_LANDING
- Took off the buffer altitude and set the touchdown speed to 1 m/s. The
script is very accurate and it will be able to land well touching down
with almost no speed.
- Changed the cuttoff conditions to better logic and fall back logic
incase it doesn't go below the buffer_alt